requirements.txt
efficientnet==1.1.1
Keras-Preprocessing==1.1.2
numpy==1.19.1
Pillow==8.0.1
streamlit==0.72.0
tensorflow-cpu==2.2.0
